# openapi.yaml
openapi: 3.0.3
info:
  title: Zakk Flask Vulnerable API
  version: 1.0.0
servers:
  - url: https://zakk-app-flask-fyeybqh6gmgdhrbr.canadacentral-01.azurewebsites.net
paths:
  /api/greeting:
    get:
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Greeting message
          content:
            text/html:
              schema:
                type: string
  /api/user:
    get:
      parameters:
        - name: id
          in: query
          schema:
            type: string   # IMPORTANT for SQLi testing
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                type: object
...